- Both 32 and 6 bit installers works correctly - All subagents ported to 64bit - Agent now reports platform windows-x64 instead of windows-amd64

+++ src/agent/core/service.cpp
@@ -302,7 +302,7 @@ void InstallEventSource(char *path)
    }
 
    RegSetValueEx(hKey, "TypesSupported", 0, REG_DWORD, (BYTE *)&dwTypes, sizeof(DWORD));
-   RegSetValueEx(hKey, "EventMessageFile", 0, REG_EXPAND_SZ, (BYTE *)path, strlen(path) + 1);
+   RegSetValueEx(hKey, "EventMessageFile", 0, REG_EXPAND_SZ, (BYTE *)path, (DWORD)(strlen(path) + 1));
 
    RegCloseKey(hKey);
    printf("Event source \"" AGENT_EVENT_SOURCE "\" installed successfully\n");
